* Renamed nvgPick/nvgPickAll to nvgHitTest/nvgHitTestAll * Renamed nvgPickFill/nvgPickStroke to nvgFillHitRegion/nvgStrokeHitRegion * Added nvgInFill/nvgInStroke * Added nanovg_internal.h for internal details. * nanovg.c no longer includes nanovg_pick.c, the latter is free-standing now. * Fixed memory leak of sub paths
