+++ premake4.lua
@@ -10,7 +10,7 @@ solution "nanovg"
 		language "C"
 		kind "StaticLib"
 		includedirs { "src" }
-		files { "src/nanovg.c" }
+		files { "src/*.c" }
 		targetdir("build")
 		defines { "_CRT_SECURE_NO_WARNINGS" } --,"FONS_USE_FREETYPE" } Uncomment to compile with FreeType support
