+++ src/nanovg.c
@@ -16,13 +16,7 @@
 // 3. This notice may not be removed or altered from any source distribution.
 //
 
-#include <stdio.h>
-#include <math.h>
-#include "nanovg.h"
-#define FONTSTASH_IMPLEMENTATION
-#include "fontstash.h"
-#define STB_IMAGE_IMPLEMENTATION
-#include "stb_image.h"
+#include "nanovg_internal.h"
 
 #ifdef _MSC_VER
 #pragma warning(disable: 4100)  // unreferenced formal parameter
@@ -31,136 +25,10 @@
 #pragma warning(disable: 4706)  // assignment within conditional expression
 #endif
 
-#define NVG_INIT_FONTIMAGE_SIZE  512
-#define NVG_MAX_FONTIMAGE_SIZE   2048
-#define NVG_MAX_FONTIMAGES       4
-
-#define NVG_INIT_COMMANDS_SIZE 256
-#define NVG_INIT_POINTS_SIZE 128
-#define NVG_INIT_PATHS_SIZE 16
-#define NVG_INIT_VERTS_SIZE 256
-#define NVG_MAX_STATES 32
-
-#define NVG_KAPPA90 0.5522847493f	// Length proportional to radius of a cubic bezier handle for 90deg arcs.
-
-#define NVG_COUNTOF(arr) (sizeof(arr) / sizeof(0[arr]))
-
-
-enum NVGcommands {
-	NVG_MOVETO = 0,
-	NVG_LINETO = 1,
-	NVG_BEZIERTO = 2,
-	NVG_CLOSE = 3,
-	NVG_WINDING = 4,
-};
-
-enum NVGpointFlags
-{
-	NVG_PT_CORNER = 0x01,
-	NVG_PT_LEFT = 0x02,
-	NVG_PT_BEVEL = 0x04,
-	NVG_PR_INNERBEVEL = 0x08,
-};
-
-struct NVGstate {
-	NVGpaint fill;
-	NVGpaint stroke;
-	float strokeWidth;
-	float miterLimit;
-	int lineJoin;
-	int lineCap;
-	float alpha;
-	float xform[6];
-	NVGscissor scissor;
-	float fontSize;
-	float letterSpacing;
-	float lineHeight;
-	float fontBlur;
-	int textAlign;
-	int fontId;
-};
-typedef struct NVGstate NVGstate;
-
-struct NVGpoint {
-	float x,y;
-	float dx, dy;
-	float len;
-	float dmx, dmy;
-	unsigned char flags;
-};
-typedef struct NVGpoint NVGpoint;
-
-struct NVGpathCache {
-	NVGpoint* points;
-	int npoints;
-	int cpoints;
-	NVGpath* paths;
-	int npaths;
-	int cpaths;
-	NVGvertex* verts;
-	int nverts;
-	int cverts;
-	float bounds[4];
-};
-typedef struct NVGpathCache NVGpathCache;
-
-typedef struct NVGpickScene NVGpickScene;
-
-struct NVGcontext {
-	NVGparams params;
-	float* commands;
-	int ccommands;
-	int ncommands;
-	float commandx, commandy;
-	NVGstate states[NVG_MAX_STATES];
-	int nstates;
-	NVGpathCache* cache;
-	float tessTol;
-	float distTol;
-	float fringeWidth;
-	float devicePxRatio;
-	struct FONScontext* fs;
-	int fontImages[NVG_MAX_FONTIMAGES];
-	int fontImageIdx;
-	int drawCallCount;
-	int fillTriCount;
-	int strokeTriCount;
-	int textTriCount;
-	NVGpickScene* pickScene;
-};
-
-static void nvg__pickBeginFrame(NVGcontext* ctx, int width, int height);
-static void nvg__pickSceneFree(NVGpickScene* ps);
-
-static float nvg__sqrtf(float a) { return sqrtf(a); }
-static float nvg__modf(float a, float b) { return fmodf(a, b); }
-static float nvg__sinf(float a) { return sinf(a); }
-static float nvg__cosf(float a) { return cosf(a); }
-static float nvg__tanf(float a) { return tanf(a); }
-static float nvg__atan2f(float a,float b) { return atan2f(a, b); }
-static float nvg__acosf(float a) { return acosf(a); }
-
-static int nvg__mini(int a, int b) { return a < b ? a : b; }
-static int nvg__maxi(int a, int b) { return a > b ? a : b; }
-static int nvg__clampi(int a, int mn, int mx) { return a < mn ? mn : (a > mx ? mx : a); }
-static float nvg__minf(float a, float b) { return a < b ? a : b; }
-static float nvg__maxf(float a, float b) { return a > b ? a : b; }
-static float nvg__absf(float a) { return a >= 0.0f ? a : -a; }
-static float nvg__signf(float a) { return a >= 0.0f ? 1.0f : -1.0f; }
-static float nvg__clampf(float a, float mn, float mx) { return a < mn ? mn : (a > mx ? mx : a); }
-static float nvg__cross(float dx0, float dy0, float dx1, float dy1) { return dx1*dy0 - dx0*dy1; }
-
-static float nvg__normalize(float *x, float* y)
-{
-	float d = nvg__sqrtf((*x)*(*x) + (*y)*(*y));
-	if (d > 1e-6f) {
-		float id = 1.0f / d;
-		*x *= id;
-		*y *= id;
-	}
-	return d;
-}
-
+#define FONTSTASH_IMPLEMENTATION
+#include "fontstash.h"
+#define STB_IMAGE_IMPLEMENTATION
+#include "stb_image.h"
 
 static void nvg__deletePathCache(NVGpathCache* c)
 {
@@ -442,7 +310,7 @@ NVGcolor nvgHSLA(float h, float s, float l, unsigned char a)
 }
 
 
-static NVGstate* nvg__getState(NVGcontext* ctx)
+NVGstate* nvg__getState(NVGcontext* ctx)
 {
 	return &ctx->states[ctx->nstates-1];
 }
@@ -2760,7 +2628,5 @@ void nvgTextMetrics(NVGcontext* ctx, float* ascender, float* descender, float* l
 	if (lineh != NULL)
 		*lineh *= invscale;
 }
-
-#include "nanovg_pick.c"
 
 // vim: ft=c nu noet ts=4
